Change col width uint to float64

+++ internal/examples/billing-with-negative/main.go
@@ -113,11 +113,11 @@ func main() {
 	m.TableList(header, contents, props.TableList{
 		HeaderProp: props.TableListContent{
 			Size:      9,
-			GridSizes: []uint{3, 4, 2, 3},
+			GridSizes: []float64{3, 4, 2, 3},
 		},
 		ContentProp: props.TableListContent{
 			Size:                            8,
-			GridSizes:                       []uint{3, 4, 2, 3},
+			GridSizes:                       []float64{3, 4, 2, 3},
 			CellTextColorChangerColumnIndex: 3,
 			CellTextColorChangerFunc: func(cellValue string) color.Color {
 				if strings.Contains(cellValue, "-R$") {
